Lecture 01 Tasks

+++ LECTURE_01/task_03.py
@@ -1,4 +1,4 @@
 #take input side of square from user and print area of square
-side = int(input("Enter side of square :"))
+side = float(input("Enter side of square :"))
 area = side * side
 print(f"Area of square with side {side} is : ", area)
